refactor : 패키지명 rapository -> repository 수정 onMessage 에서 문자열 분한 로직 중복 수정
