+++ backend/src/main/java/com/project/backend/global/redis/service/RedisPublisher.java
@@ -1,6 +1,7 @@
 package com.project.backend.global.redis.service;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Service;
